fix(session): drain poller observations in daemon-mode status loop
Refs #2291

Sessions running under `aoe serve` without an attached TUI did not
persist post-`/clear` session ids: the TUI was the only consumer of
the mpsc that `SessionPoller::on_change` pushes to, so `sessions.json`
stayed stale until the next launch's resume-time rescan. The Claude
sidecar fallback closed most of this for Claude alone; the other six
live-tracked tools (opencode, vibe, codex, gemini, pi, hermes) had no
fallback at all in daemon-only mode.

Lift the drain + persist + tmux-env-publish path out of
`tui/home/mod.rs::apply_session_id_updates` into a shared
`src/session/sync.rs::drain_and_persist_session_ids`. The TUI keeps
its public function signature and re-syncs `instance_map` from the
mutated slice; the daemon's `status_poll_loop` calls the same helper
on `state.instances` after `reload_state_instances_from_disk`, on
the existing 2s tick.

Each instance's `Arc<Mutex<SessionPoller>>` is preserved across
daemon disk reloads via `merge_runtime_fields`, and `try_recv` is
exclusive so a TUI-attached daemon's two drain sites cannot double-
apply. Persist goes through the same `persist_session_to_storage`
CAS, so a loser hits Skipped and reloads from disk.
